+++ contrib/scoverage/src/mill/contrib/scoverage/ScoverageModule.scala
@@ -7,7 +7,7 @@ import mill.contrib.scoverage.api.ScoverageReportWorkerApi.ReportType
 import mill.scalalib.api.ZincWorkerUtil
 import mill.scalalib.{Dep, DepSyntax, JavaModule, ScalaModule}
 import mill.api.Result
-import mill.modules.Util.millProjectModule
+import mill.util.Util.millProjectModule
 
 import scala.util.Try
