+++ scalalib/test/src/mill/scalalib/HelloWorldTests.scala
@@ -9,7 +9,7 @@ import mill._
 import mill.api.Result
 import mill.define.NamedTask
 import mill.eval.{Evaluator, EvaluatorPaths}
-import mill.modules.Assembly
+import mill.scalalib.Assembly
 import mill.scalalib.api.ZincWorkerUtil
 import mill.scalalib.publish.{VersionControl, _}
 import mill.util.{TestEvaluator, TestUtil}
