+++ scalanativelib/src/mill/scalanativelib/ScalaNativeModule.scala
@@ -5,8 +5,8 @@ import mainargs.Flag
 import mill.api.Loose.Agg
 import mill.api.{Result, internal}
 import mill.define.{Command, Target, Task}
-import mill.modules.Jvm
-import mill.modules.Util.millProjectModule
+import mill.util.Jvm
+import mill.util.Util.millProjectModule
 import mill.scalalib.api.ZincWorkerUtil
 import mill.scalalib.bsp.{ScalaBuildTarget, ScalaPlatform}
 import mill.scalalib.{
